Basic data sending

+++ SDCard.cpp
@@ -104,7 +104,7 @@ void SDCard::checkCardMemory()
 /*
  * 
  */
-void SDCard::printCurrentMeasurement(int measurement)
+void SDCard::printCurrentMeasurement(int16_t measurement)
 {
   char currentMeasurementMessage[] PROGMEM = "Current Measurement: ";
   Serial.println(currentMeasurementMessage);
@@ -115,7 +115,7 @@ void SDCard::printCurrentMeasurement(int measurement)
 /*
  * 
  */
-void SDCard::printToLog(int lastMeasurementSent)
+void SDCard::printToLog(int16_t lastMeasurementSent)
 {
