+++ Sensor.h
@@ -6,21 +6,23 @@
 #ifndef Sensor_h
 #define Sensor_h
 
+#include "Arduino.h"
+
 /**
  * 
  */
 class Sensor
 {
   public:
-    int currentRiverLevel;
-    int distanceToRiverBed;
-    int rangeDifferenceThreshold;
-    int lastMeasurementSent;
-    int analogPin;
+    int16_t currentRiverLevel;
+    int16_t distanceToRiverBed;
+    int16_t rangeDifferenceThreshold;
+    int16_t lastMeasurementSent;
+    int16_t analogPin;
   
-    Sensor(int analogPin);
-    int getCurrentMeasurement();
-    bool isCurrentWorthSending(int currentMeasurement);
+    Sensor(uint8_t analogPin);
+    int16_t getCurrentMeasurement();
+    bool isCurrentWorthSending(int16_t currentMeasurement);
   private:
 
 };
